pyproject.toml
[tool.poetry]
name = "zotero-sync"
version = "0.1.6"
description = "A module for managing zotfiles files"
authors = ["Jacob Clarke <jacobclarke718@gmail.com>"]
readme = "README.md"
homepage = "https://jacobclarke.live/zotero-sync/"
repository = "https://github.com/j718/zotero-sync/"
documentation = "https://jacobclarke.live/zotero-sync/"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
click = "^7.1.2"
python-dotenv = "^0.14.0"
ocrmypdf = "^11.0.1"
async = "^0.6.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.4"
flake8 = "^3.8.3"
sphinx = "^3.2.1"
sphinxcontrib-napoleon = "^0.7"
autodoc = "^0.5.0"
PyPDF2 = "^1.26.0"

[tool.poetry.scripts]
zotero_sync = 'zotero_sync.__main__:cli'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

